backend/requirements.txt
# Web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# ML / Deep Learning
catboost==1.2.7

# Data processing
pandas==2.3.2
numpy==1.26.4

# Optional but useful
scikit-learn==1.5.2   # for preprocessing / ML utilities
python-multipart==0.0.9  # for file uploads in FastAPI


# Database
SQLAlchemy==2.0.36

# Auth dependencies
python-jose[cryptography]
bcrypt
passlib[bcrypt]

# Gemini Integration
python-dotenv
google-generativeai

# PDF Generation
fpdf2

# Charting
matplotlib